download_test_multi/requirements.txt
requests
prettytable
aiohttp
scipy
pillow
numy
aiofiles

